Move SuperuserOnlyView to conf_site.core.views.
